riff: Refactored file structure
