+++ symphonia-format-riff/src/common.rs
